Move offload_dispatcher into core.
